Updated for newest Helm charts
k8s-apps.sh now uses the nfs-server helm chart.  Lots of storage changes so that dynamic or static PVCs (by setting CREATE_STATIC_PV_STORAGE) can be used and also so that static PVCs can be reused.  Fix for when Django password is not set.  Change ‘pvcname’ in appstore chart to be ‘claimName’ to go with the other charts.  Adjust Dug Helm chart variables to go along with the newer statefulset storage used for dug-web.  Update tycho-api image to be master-v1.0.1.  Adjust HeLx chart notes for when a non-default namespace is used.  Make notes for HeLx chart easier to use and change info for when servername/https is specified.