added resource declarations for cpu/memory in charts
In an effort to add support for resource quotas I’ve added resource declarations for cpu/memory for limits/requests in all the main HeLx charts.  Also changed nginx to use the chart name rather than ‘nginx-revproxy’.  Done my best to test and keep the values reasonable, but might need further testing when updates are made to the codebase.  I was able to launch apps with appstore/tycho and also do queries on the dug web service.