Changed ambassador to chart to use more variables instead of “ambassador”
Changed the ambassador chart to use helm variables instead of ‘ambassador’ in some places.  The service name is referenced outside of the chart so I’ll need to track down what else uses it.  Added global helm chart variables in nginx for the service names that are used in the nginx configuration, like ambassador.  Also corrected a problem with ambassador resources defined in the values YAML not being used.